Re-add rest server test
The original intention was to use Cucumber to test the entire rest
server, but it isn't useful to use BDD for this task.
